+++ NEWS.md
@@ -1,5 +1,7 @@
 # quarto (development version)
 
+- Add `quarto_available()` function to check if Quarto CLI is found (thanks, @hadley, #187).
+
 - Add several new wrapper function (thanks, @parmsam, #192): 
   - `quarto_list_extensions()` to list installed extensions using `quarto list extensions`
   - `quarto_remove_extension()` to remove an installed extension using `quarto remove extensions`